// src/types/supabase.ts
export type Database = {
    // Allows to automatically instantiate createClient with right options
    // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
    __InternalSupabase: {
        PostgrestVersion: "14.1"
    }
    public: {
        Tables: {
// ...
            site_stats: {
                Row: {
                    created_at: string | null
                    display_order: number | null
                    id: string
                    key: string
                    label_ar: string | null
                    label_en: string | null
                    subtitle_ar: string | null
                    subtitle_en: string | null
                    suffix: string | null
                    value: number | null
                }
                Insert: {
                    created_at?: string | null
                    display_order?: number | null
                    id?: string
                    key: string
                    label_ar?: string | null
                    label_en?: string | null
                    subtitle_ar?: string | null
                    subtitle_en?: string | null
                    suffix?: string | null
                    value?: number | null
                }
                Update: {
                    created_at?: string | null
                    display_order?: number | null
                    id?: string
                    key?: string
                    label_ar?: string | null
                    label_en?: string | null
                    subtitle_ar?: string | null
                    subtitle_en?: string | null
                    suffix?: string | null
                    value?: number | null
                }
                Relationships: []
            }
// ...
        }
        Views: {
            [_ in never]: never
        }
        Functions: {
            [_ in never]: never
        }
        Enums: {
            [_ in never]: never
        }
        CompositeTypes: {
            [_ in never]: never
        }
    }
}